Server_VPS/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
paho-mqtt==2.0.0
websockets==12.0
aiofiles==23.2.1